source file upload -- src (third party modules)
third party modules reproduced to use toolz as the main source of
dependencies. the original project (which was to be a readme file
generator) had only these files. the initial goal was to use two
dependencies: handlebars and toolz. toolz needed highly visible
projects to get a partial run through and ensure consistency while
working out bugs. My apologies to the authors.