update index.js
stopped deleting template cache. instead I update the cache with the
file that changed. no longer is it necessary to run loadTemplates after
a rendr. removed buildFileTree as well. It was only being used in
testing.

Also, file that changes will be added to file map cache only if it
resides in the `views` directory.