update index.js
increase the speed up further by only passing the path of the changed
layout or template for buildLayoutStack and frontMatter respectively.
Now do not have to iterate over all the files since we are
updating/returning the cache.

still using globby.sync…this ensures an array without have to put a
check in the function